+++ Services/CoreLibrary.cs
@@ -34,7 +34,6 @@ public class CoreLibrary
     public WebhookService Webhook { get; }
     public FileWatcherService FileWatcher { get; }
     public Action<string, object?> SendToJS { get; }
-    public Action<string>? AvtrdbSubmit { get; set; }
     public Action<string>? VrcndbSubmit { get; set; }
 
     public ConcurrentDictionary<string, bool> PlayerAgeVerifiedCache { get; } = new();